requirements.txt
python-dotenv>=0.19.0
beautifulsoup4>=4.9.3
playwright>=1.40.0
openai>=1.0.0
pandas>=2.0.0
newspaper3k>=0.2.8
requests>=2.26.0
feedparser>=6.0.0
flask>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
langgraph>=0.0.15
langchain>=0.1.0
langchain-openai>=0.0.2
pytest>=7.0.0
pytest-asyncio>=0.21.0
google-adk==1.1.1 
fastapi==0.115.12 
uvicorn[standard]==0.34.2 
fastmcp==2.5.1 
python-dotenv==1.1.0
google-api-python-client
youtube-transcript-api
pandas
